+++ trunk/JPAS/src/org/jpas/model/ReminderTransfer.java
@@ -24,50 +24,26 @@
 package org.jpas.model;
 
 import org.jpas.da.ReminderAccountMappingDA;
-import org.jpas.util.JpasDataChange;
-import org.jpas.util.JpasObservable;
-import org.jpas.util.WeakValueMap;
+import org.jpas.da.TransAccountMappingDA;
+import org.jpas.util.*;
 
 /**
  * @author Justin W Smith
  *  
  */
-public class ReminderTransfer extends JpasObservable<ReminderTransfer>
+public class ReminderTransfer extends JpasObservableImpl
 {
-    private static WeakValueMap<Integer, WeakValueMap<Integer, ReminderTransfer>> remTransferCache = new WeakValueMap<Integer, WeakValueMap<Integer, ReminderTransfer>>();
-    private static JpasObservable<ReminderTransfer> observable = new JpasObservable<ReminderTransfer>();
-
-    public static JpasObservable<ReminderTransfer> getObservable()
-    {
-        return observable;
-    }
-
-    private boolean isDeleted = false;
     private boolean isLoaded = false;
+    private boolean isDeleted = false;
+    private boolean isModified = false;
+    
     final Integer reminderID;
-    Integer accountID;
+    final Integer accountID;
     private long amount;
 
-    static ReminderTransfer getReminderTransferforIDs(final Integer reminderID,
+    
+    ReminderTransfer(final Integer reminderID,
             final Integer accountID)
-    {
-        WeakValueMap<Integer, ReminderTransfer> map = remTransferCache
-                .get(reminderID);
-        if (map == null)
-        {
-            map = new WeakValueMap<Integer, ReminderTransfer>();
-            remTransferCache.put(reminderID, map);
-        }
-        ReminderTransfer rt = map.get(accountID);
-        if (rt == null)
-        {
-            rt = new ReminderTransfer(reminderID, accountID);
-            map.put(accountID, rt);
-        }
-        return rt;
-    }
-
-    private ReminderTransfer(final Integer reminderID, final Integer accountID)
     {
         this.reminderID = reminderID;
         this.accountID = accountID;
@@ -75,20 +51,20 @@ public class ReminderTransfer extends JpasObservable<ReminderTransfer>
 
     public Category getCategory()
     {
-        return Category.getCategoryForID(accountID);
+        return ModelFactory.getInstance().getAccountImplForID(accountID);
     }
 
-    public Reminder getTransaction()
+    public Reminder getReminder()
     {
-        return Reminder.getReminderForID(reminderID);
+        return ModelFactory.getInstance().getReminderForID(reminderID);
     }
 
     private void loadData()
     {
         assert (!isDeleted);
-        ReminderAccountMappingDA.getInstance().loadReminderAccountMapping(
+        TransAccountMappingDA.getInstance().loadTransAccountMapping(
                 reminderID, accountID,
-                new ReminderAccountMappingDA.ReminderAccountTranferHandler()
+                new TransAccountMappingDA.TransAccountTranferHandler()
                 {
                     public void setData(final long amount)
                     {
@@ -107,52 +83,56 @@ public class ReminderTransfer extends JpasObservable<ReminderTransfer>
         return amount;
     }
 
+    private void announceChange(final JpasDataChange change)
+    {
+    	notifyObservers(change);
+    }
+    
+    public void commit()
+    {
+    	if(isModified)
+    	{
+    		if(isDeleted)
+    		{
+	    		final JpasDataChange change = new JpasDataChange.Delete(this);
+	    		TransAccountMappingDA.getInstance().deleteTransAccountMapping(reminderID, accountID);
+	    		announceChange(change);
+	    		deleteObservers();
+	    		isModified = false;
+	    	}
+	    	else
+	    	{
+	    		final JpasDataChange change = new JpasDataChange.AmountModify(this);
+	    		TransAccountMappingDA.getInstance().updateTAMAmount(reminderID, accountID, amount);
+	    		announceChange(change);
+	    		isModified = false;
+	    	}
+    	}
+    }
+
     public void setAmount(final long amount)
     {
         assert (!isDeleted);
-        ReminderAccountMappingDA.getInstance().updateReminderAccountMapping(
-                reminderID, accountID, amount);
         if (isLoaded)
         {
             loadData();
         }
-        announceModify();
-        Reminder.getReminderForID(reminderID).amountChanged();
+        if(this.amount != amount)
+        {
+        	isModified = true;
+        	this.amount = amount;
+        }
     }
 
-    public void setCategory(final Category category)
+    public void delete()
     {
-        assert (!isDeleted);
-        ReminderAccountMappingDA.getInstance().updateReminderAccountMapping(
-                reminderID, accountID, amount);
-        if (isLoaded)
+        if(!isDeleted)
         {
-            loadData();
+        	isDeleted = true;
+        	isModified = true;
         }
-        
-        announceModify();
-        category.amountChanged();
-        Reminder.getReminderForID(reminderID).amountChanged();
     }
-
     
-    private void announceDelete()
-    {
-        final JpasDataChange<ReminderTransfer> change = new JpasDataChange.Delete<ReminderTransfer>(
-                this);
-        observable.notifyObservers(change);
-        notifyObservers(change);
-        deleteObservers();
-    }
-
-    private void announceModify()
-    {
-        final JpasDataChange<ReminderTransfer> change = new JpasDataChange.Modify<ReminderTransfer>(
-                this);
-        observable.notifyObservers(change);
-        notifyObservers(change);
-    }
-
     public boolean isDeleted()
     {
         return isDeleted;
@@ -163,35 +143,13 @@ public class ReminderTransfer extends JpasObservable<ReminderTransfer>
         return isLoaded;
     }
 
-    public void delete()
+    public boolean isModified()
     {
-        delete(false);
+    	return isModified;
     }
-
-    void delete(final boolean internalCall)
-    {
-        if (!internalCall)
-        {
-            ReminderAccountMappingDA.getInstance()
-                    .deleteReminderAccountMapping(reminderID, accountID);
-        }
-        final WeakValueMap<Integer, ReminderTransfer> map = remTransferCache
-                .get(reminderID);
-        map.remove(accountID);
-        if (map.size() == 0)
-        {
-            remTransferCache.remove(reminderID);
-        }
-        isDeleted = true;
-        
-        announceDelete();
-        if (!internalCall)
-        {
-            Reminder.getReminderForID(reminderID).amountChanged();
-        }
-    }
-
+    
     public static void main(String[] args)
     {
     }
+
 }
